tag-tools/youtube/requirements.txt
google-api-python-client
tenacity 
ratelimit
python-dotenv